+++ doc/releases/changelog-dev.md
@@ -110,6 +110,8 @@
 
 <h3>Improvements 🛠</h3>
 
+* `qml.to_openqasm` now supports mid circuit measurements and conditionals of unprocessed measurement values.
+
 * `allocate` and `deallocate` can now be accessed as `qml.allocate` and `qml.deallocate`.
   [(#8189)](https://github.com/PennyLaneAI/pennylane/pull/8198))
