+++ tests/io/test_qasm.py
@@ -33,7 +33,7 @@ class TestToQasmUnitTests:
         serialized into an empty QASM program."""
         circuit = qml.tape.QuantumScript()
         res = qml.to_openqasm(circuit)
-        expected = 'OPENQASM 2.0;\ninclude "qelib1.inc";\n'
+        expected = 'OPENQASM 2.0;\ninclude "qelib1.inc";'
         assert res == expected
 
     def test_native_qasm_gates(self):
@@ -66,8 +66,7 @@ class TestToQasmUnitTests:
             x q[1];
             measure q[0] -> c[0];
             measure q[1] -> c[1];
-            measure q[2] -> c[2];
-            """
+            measure q[2] -> c[2];"""
         )
 
         assert res == expected
@@ -93,8 +92,7 @@ class TestToQasmUnitTests:
             measure q[1] -> c[1];
             measure q[2] -> c[2];
             measure q[3] -> c[3];
-            measure q[4] -> c[4];
-            """
+            measure q[4] -> c[4];"""
         )
 
         assert res == expected
@@ -116,8 +114,7 @@ class TestToQasmUnitTests:
             rz(2.0) q[0];
             cx q[1],q[0];
             measure q[0] -> c[0];
-            measure q[1] -> c[1];
-            """
+            measure q[1] -> c[1];"""
         )
 
         assert res == expected
@@ -147,8 +144,7 @@ class TestToQasmUnitTests:
             ry(0.1) q[1];
             rz(0.2) q[1];
             measure q[0] -> c[0];
-            measure q[1] -> c[1];
-            """
+            measure q[1] -> c[1];"""
         )
 
         assert qasm1 == expected
@@ -185,8 +181,7 @@ class TestToQasmUnitTests:
             rz(3.1415926536) q[1];
             cx q[0],q[1];
             measure q[0] -> c[0];
-            measure q[1] -> c[1];
-            """
+            measure q[1] -> c[1];"""
         )
 
         # different
@@ -222,8 +217,7 @@ class TestToQasmUnitTests:
             measure q[0] -> c[0];
             measure q[1] -> c[1];
             measure q[2] -> c[2];
-            measure q[3] -> c[3];
-            """
+            measure q[3] -> c[3];"""
         )
 
         assert qasm1 == expected
@@ -254,8 +248,7 @@ class TestToQasmUnitTests:
             ry(-0.7853981633974483) q[2];
             measure q[0] -> c[0];
             measure q[1] -> c[1];
-            measure q[2] -> c[2];
-            """
+            measure q[2] -> c[2];"""
         )
 
         assert res == expected
@@ -301,8 +294,7 @@ class TestToQasmUnitTests:
             cx q[2],q[0];
             x q[1];
             measure q[0] -> c[0];
-            measure q[2] -> c[2];
-            """
+            measure q[2] -> c[2];"""
         )
 
         assert res == expected
@@ -324,8 +316,7 @@ class TestQNodeQasmIntegrationTests:
             include "qelib1.inc";
             qreg q[1];
             creg c[1];
-            measure q[0] -> c[0];
-            """
+            measure q[0] -> c[0];"""
         )
         assert res == expected
 
@@ -361,8 +352,7 @@ class TestQNodeQasmIntegrationTests:
             x q[1];
             measure q[0] -> c[0];
             measure q[1] -> c[1];
-            measure q[2] -> c[2];
-            """
+            measure q[2] -> c[2];"""
         )
 
         assert res == expected
@@ -405,8 +395,7 @@ class TestQNodeQasmIntegrationTests:
             x q[1];
             measure q[0] -> c[0];
             measure q[1] -> c[1];
-            measure q[2] -> c[2];
-            """
+            measure q[2] -> c[2];"""
         )
 
         assert res == expected
@@ -431,8 +420,7 @@ class TestQNodeQasmIntegrationTests:
             x q[1];
             measure q[0] -> c[0];
             measure q[1] -> c[1];
-            measure q[2] -> c[2];
-            """
+            measure q[2] -> c[2];"""
         )
 
         assert res == expected
@@ -472,8 +460,7 @@ class TestQNodeQasmIntegrationTests:
             measure q[1] -> c[1];
             measure q[2] -> c[2];
             measure q[3] -> c[3];
-            measure q[4] -> c[4];
-            """
+            measure q[4] -> c[4];"""
         )
 
         assert res == expected
@@ -498,8 +485,7 @@ class TestQNodeQasmIntegrationTests:
             ry(0.1) q[1];
             rz(0.2) q[1];
             measure q[0] -> c[0];
-            measure q[1] -> c[1];
-            """
+            measure q[1] -> c[1];"""
         )
 
         assert res == expected
@@ -532,8 +518,7 @@ class TestQNodeQasmIntegrationTests:
             rz(3.1415926536) q[1];
             cx q[0],q[1];
             measure q[0] -> c[0];
-            measure q[1] -> c[1];
-            """
+            measure q[1] -> c[1];"""
         )
 
         assert res == expected
@@ -564,8 +549,7 @@ class TestQNodeQasmIntegrationTests:
             measure q[0] -> c[0];
             measure q[1] -> c[1];
             measure q[2] -> c[2];
-            measure q[3] -> c[3];
-            """
+            measure q[3] -> c[3];"""
         )
 
         assert res == expected
@@ -594,8 +578,7 @@ class TestQNodeQasmIntegrationTests:
             ry(-0.7853981633974483) q[2];
             measure q[0] -> c[0];
             measure q[1] -> c[1];
-            measure q[2] -> c[2];
-            """
+            measure q[2] -> c[2];"""
         )
 
         assert res == expected
@@ -628,8 +611,7 @@ class TestQNodeQasmIntegrationTests:
             ry(-0.7853981633974483) q[1];
             measure q[0] -> c[0];
             measure q[1] -> c[1];
-            measure q[2] -> c[2];
-            """
+            measure q[2] -> c[2];"""
         )
 
         assert res == expected
@@ -646,8 +628,7 @@ class TestQNodeQasmIntegrationTests:
             qreg q[1];
             creg c[1];
             rx(3.142) q[0];
-            measure q[0] -> c[0];
-            """
+            measure q[0] -> c[0];"""
         )
 
         assert res == expected
@@ -674,10 +655,82 @@ class TestQNodeQasmIntegrationTests:
             qreg q[1];
             creg c[1];
             rx(1.2) q[0];
-            measure q[0] -> c[0];
-            """
+            measure q[0] -> c[0];"""
         )
 
+        assert res == expected
+
+    def test_error_reset_True(self):
+        """Test an error is raised if the mcm has reset"""
+        m0 = qml.measure(0, reset=True)
+        tape = qml.tape.QuantumScript(m0.measurements)
+        with pytest.raises(NotImplementedError):
+            qml.to_openqasm(tape)
+
+    def test_error_postselection(self):
+        """Test that an error is raised if postselection exists."""
+        m0 = qml.measure(0, postselect=1)
+        tape = qml.tape.QuantumScript(m0.measurements)
+        with pytest.raises(NotImplementedError):
+            qml.to_openqasm(tape)
+
+    def test_error_if_mcm_processed(self):
+        """Test a NotImplementedError is raised if the mcm is processed in the conditional."""
+        m0 = qml.measure(0)
+        tape = qml.tape.QuantumScript([m0.measurements[0], qml.ops.Conditional(2 * m0, qml.X(0))])
+        with pytest.raises(NotImplementedError):
+            qml.to_openqasm(tape)
+
+    def test_multiple_mcms(self):
+        """Test that multiple mcms can be translated."""
+
+        m0 = qml.measure(0)
+        m1 = qml.measure(1)
+        m2 = qml.measure(2)
+
+        tape = qml.tape.QuantumScript(
+            [qml.X(0), *m0.measurements, *m1.measurements, *m2.measurements]
+        )
+
+        expected = dedent(
+            """\
+                            OPENQASM 2.0;
+                            include "qelib1.inc";
+                            qreg q[3];
+                            creg c[3];
+                            creg mcms[3]
+                            x q[0];
+                            measure q[0] -> mcms[0];
+                            measure q[1] -> mcms[1];
+                            measure q[2] -> mcms[2];
+                            measure q[0] -> c[0];
+                            measure q[1] -> c[1];
+                            measure q[2] -> c[2];"""
+        )
+        assert expected == qml.to_openqasm(tape)
+
+    @pytest.mark.parametrize("precision", (None, 2))
+    def test_conditional(self, precision):
+        """Test that a conditional can be translated."""
+
+        m0 = qml.measure(0)
+        tape = qml.tape.QuantumScript(
+            [m0.measurements[0], qml.ops.Conditional(m0, qml.RX(0.123456, 0))]
+        )
+        res = qml.to_openqasm(tape, precision=precision)
+
+        p = f"{0.123456:.{precision}}" if precision else str(0.123456)
+        expected = dedent(
+            f"""\
+                    OPENQASM 2.0;
+                    include "qelib1.inc";
+                    qreg q[1];
+                    creg c[1];
+                    creg mcms[1]
+                    measure q[0] -> mcms[0];
+                    if(mcms[0]==1) rx({p}) q[0];
+                    measure q[0] -> c[0];"""
+        )
         assert res == expected
